+++ p2panda-rs/src/schema/mod.rs
@@ -16,7 +16,7 @@ mod schema;
 
 pub use error::SchemaError;
 pub use operation::OPERATION_SCHEMA;
-pub use schema::{Schema, SchemaBuilder, Type};
+pub use schema::{Schema, SchemaBuilder, Type, ValidateOperation};
 
 /// Checks CBOR bytes against CDDL schemas.
 ///